server: split taskIngest.go
