+++ server/taskIngest.go
@@ -47,31 +47,6 @@ func taskIngest(ctx context.Context) error {
 	}
 }
 
-func taskIngestDue(ctx context.Context) (err error) {
-	var tx *sql.Tx
-	logger := jsonlog.LoggerFromContextOrDefault(ctx)
-	defer func() {
-		if tx != nil {
-			if err != nil {
-				tx.Rollback()
-				logger.Debug("Rolled back transaction.", nil)
-			} else {
-				tx.Commit()
-				logger.Debug("Commited transaction.", nil)
-			}
-		}
-	}()
-	if tx, err = db.Db.BeginTx(ctx, nil); err != nil {
-	} else {
-		logger.Debug("Started transaction.", nil)
-		conclusion, err := s3.UpdateDueReports(ctx, tx)
-		if err == nil {
-			err = updateBillRepositoriesFromConclusion(ctx, tx, conclusion)
-		}
-	}
-	return
-}
-
 func updateBillRepositoriesFromConclusion(ctx context.Context, tx *sql.Tx, ruccs []s3.ReportUpdateConclusion) error {
 	for _, r := range ruccs {
 		if r.Error != nil {
